refactor(pywire): move validate_project from cli to compiler
Breaks core → cli circular dep. Also fixes glob bug (.pywire → .wire).

Prep for splitting pywire-cli into own package.
