+++ packages/pywire/src/pywire/compiler/validate.py
@@ -1,4 +1,4 @@
-"""Validation for .pywire files."""
+"""Validation for .wire files."""
 
 from pathlib import Path
 from typing import List
@@ -7,20 +7,19 @@ from pywire.compiler.parser import PyWireParser
 
 
 def validate_project(pages_dir: Path) -> List[str]:
-    """Validate all .pywire files in project."""
-    errors = []
+    """Validate all .wire files in project."""
+    errors: List[str] = []
     parser = PyWireParser()
 
     if not pages_dir.exists():
         return [f"Pages directory not found: {pages_dir}"]
 
-    for pywire_file in pages_dir.rglob("*.pywire"):
+    for wire_file in pages_dir.rglob("*.wire"):
         try:
-            parsed = parser.parse_file(pywire_file)
-            # Basic validation
+            parsed = parser.parse_file(wire_file)
             if not parsed.template and not parsed.directives:
-                errors.append(f"{pywire_file}: No template or directives found")
+                errors.append(f"{wire_file}: No template or directives found")
         except Exception as e:
-            errors.append(f"{pywire_file}: {str(e)}")
+            errors.append(f"{wire_file}: {str(e)}")
 
     return errors
